Add support for i2c and spi cirque
