+++ quantum/pointing_device.h
@@ -29,8 +29,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #elif defined(POINTING_DEVICE_DRIVER_analog_joystick)
 #    include "analog.h"
 #    include "drivers/sensors/analog_joystick.h"
-#elif defined(POINTING_DEVICE_DRIVER_cirque_pinnacle)
-#    include "i2c_master.h"
+#elif defined(POINTING_DEVICE_DRIVER_cirque_pinnacle_i2c) || defined(POINTING_DEVICE_DRIVER_cirque_pinnacle_spi)
 #    include "drivers/sensors/cirque_pinnacle.h"
 #elif defined(POINTING_DEVICE_DRIVER_pimoroni_trackball)
 #    include "i2c_master.h"
